- Removed unused methods & variables
